修复了GC阈值不能正常增长的问题

+++ vm/vm.h
@@ -38,7 +38,7 @@ typedef struct
 typedef struct
 {
     // 堆增长因子
-    int heapGrowthFactor;
+    double heapGrowthFactor;
 
     // 初始堆大小,默认为10MB
     uint32_t initialHeapSize;
